+++ packages/file-storage/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This is the changelog for [`file-storage`](https://github.com/mjackson/remix-the-web/tree/main/packages/file-storage). It follows [semantic versioning](https://semver.org/).
 
-## HEAD
+## v0.6.0 (2025-02-04)
 
 - BREAKING CHANGE: `LocalFileStorage` now uses 2 characters for shard directory names instead of 8.
 - Buffer contents of files stored in `MemoryFileStorage`.